Add files via upload
Celebrate the arrival of the New Year with this simple HTML and CSS project. This web page beautifully represents the transition from 2023 to 2024, creating a festive and welcoming atmosphere. The minimalistic design and smooth transition make it a perfect addition to your New Year's Eve celebration. Feel free to use and customize this code for your own New Year-themed web projects